Split server into modules
