+++ server.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 """remote_bot_server"""
 import asyncio
-import io
 import json
 import logging
 import os
@@ -18,20 +17,16 @@ import subprocess
 import sys
 import threading
 from pathlib import Path
-import numpy as np
-from statistics import median
 from datetime import datetime, timedelta
 from typing import Any, Dict, List
 
 import matplotlib
-import matplotlib.pyplot as plt
 import uvicorn
 from fastapi import FastAPI, HTTPException
 from pydantic import BaseModel
 from telegram import (
     InlineKeyboardButton,
     InlineKeyboardMarkup,
-    InputMediaPhoto,
     Update,
 )
 from telegram.ext import (
@@ -42,10 +37,11 @@ from telegram.ext import (
     ContextTypes,
 )
 
+from db import sql, load_db, save_db, record_metric, purge_old_metrics
+from graphs import parse_timespan, plot_custom, plot_metric, plot_net, plot_all_metrics
+
 # ────────────────────────── CONFIG ─────────────────────────────────────────
 ENV_FILE = Path(".env")
-DB_FILE = Path("db.json")
-METRIC_DB = Path("metrics.sqlite")
 API_PORT = int(os.getenv("PORT", "8000"))
 
 # последний текстовый статус от клиентов (speedtest и пр.)
@@ -189,115 +185,6 @@ async def check_speedtest_done(ctx: ContextTypes.DEFAULT_TYPE):
     )
     job.schedule_removal()
 
-# ───────────────────── SQLite helpers ──────────────────────────────────────
-def _init_metric_db() -> sqlite3.Connection:
-    con = sqlite3.connect(METRIC_DB, check_same_thread=False, isolation_level=None)
-    con.row_factory = sqlite3.Row
-    con.execute(
-        """CREATE TABLE IF NOT EXISTS metrics(
-               secret TEXT,
-               ts     INTEGER,
-               cpu    REAL,
-               ram    REAL,
-               gpu    REAL,
-               vram   REAL,
-               ram_used   REAL,
-               ram_total  REAL,
-               swap       REAL,
-               swap_used  REAL,
-               swap_total REAL,
-               vram_used  REAL,
-               vram_total REAL,
-               cpu_temp   REAL,
-               gpu_temp   REAL,
-               net_up     REAL,
-               net_down   REAL,
-               uptime     INTEGER,
-               disks      TEXT
-        )"""
-    )
-    cols = [r[1] for r in con.execute("PRAGMA table_info(metrics)")]
-    if "net_up" not in cols:
-        con.execute("ALTER TABLE metrics ADD COLUMN net_up REAL")
-    if "net_down" not in cols:
-        con.execute("ALTER TABLE metrics ADD COLUMN net_down REAL")
-    con.execute(
-        "CREATE INDEX IF NOT EXISTS idx_metrics_secret_ts ON metrics(secret, ts)"
-    )
-    return con
-
-sql = _init_metric_db()
-
-def purge_old_metrics(days: int = 30):
-    """Удалить метрики старше указанного количества дней."""
-    cutoff = int(time.time()) - days * 86400
-    cur = sql.execute("DELETE FROM metrics WHERE ts < ?", (cutoff,))
-    log.info("🧹 Purged %d old metric rows", cur.rowcount)
-
-def record_metric(secret: str, data: Dict[str, Any]):
-    sql.execute(
-        """INSERT INTO metrics(
-               secret, ts, cpu, ram, gpu, vram,
-               ram_used, ram_total, swap, swap_used, swap_total,
-               vram_used, vram_total, cpu_temp, gpu_temp,
-               net_up, net_down, uptime, disks
-           ) VALUES(?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)""",
-        (
-            secret,
-            int(time.time()),
-            data.get("cpu"),
-            data.get("ram"),
-            data.get("gpu"),
-            data.get("vram"),
-            data.get("ram_used"),
-            data.get("ram_total"),
-            data.get("swap"),
-            data.get("swap_used"),
-            data.get("swap_total"),
-            data.get("vram_used"),
-            data.get("vram_total"),
-            data.get("cpu_temp"),
-            data.get("gpu_temp"),
-            data.get("net_up"),
-            data.get("net_down"),
-            data.get("uptime"),
-            json.dumps(data.get("disks")),
-        ),
-    )
-
-def fetch_metrics(secret: str, since: int) -> List[tuple[int, float]]:
-    rows = sql.execute(
-        "SELECT ts, cpu, ram, gpu, vram, net_up, net_down FROM metrics WHERE secret=? AND ts>=? ORDER BY ts ASC",
-        (secret, since),
-    ).fetchall()
-
-    if not rows:
-        return []
-
-    grouped = []
-    chunk: list[sqlite3.Row] = []
-    for r in rows:
-        chunk.append(r)
-        if len(chunk) == 6:
-            grouped.append(_avg_chunk(chunk))
-            chunk = []
-    if chunk:
-        grouped.append(_avg_chunk(chunk))
-    return grouped
-
-def _avg(val_list: List[float | None]) -> float | None:
-    vals = [v for v in val_list if v is not None]
-    return sum(vals) / len(vals) if vals else None
-
-def _avg_chunk(chunk: List[sqlite3.Row]) -> tuple[int, float | None, float | None, float | None, float | None, float | None, float | None]:
-    ts = chunk[-1][0]
-    cpu = _avg([r[1] for r in chunk])
-    ram = _avg([r[2] for r in chunk])
-    gpu = _avg([r[3] for r in chunk])
-    vram = _avg([r[4] for r in chunk])
-    up = _avg([r[5] for r in chunk])
-    down = _avg([r[6] for r in chunk])
-    return ts, cpu, ram, gpu, vram, up, down
 
 
 async def _purge_loop():
@@ -332,20 +219,6 @@ async def maybe_send_alerts(secret: str, data: Dict[str, Any]):
     if changed:
         save_db(db)
 
-# ──────────────────────── JSON DB helpers ──────────────────────────────────
-def load_db() -> Dict[str, Any]:
-    if DB_FILE.exists():
-        data = json.loads(DB_FILE.read_text())
-    else:
-        data = {}
-    data.setdefault("secrets", {})
-    data.setdefault("active", {})
-    data.setdefault("alerts", {})
-    data.setdefault("alert_last", {})
-    return data
-
-def save_db(db: Dict[str, Any]):
-    DB_FILE.write_text(json.dumps(db, indent=2))
 
 # ───────────────────────- Telegram command handlers ────────────────────────
 OWNER_HELP = (
@@ -770,395 +643,6 @@ async def cmd_plot(update: Update, ctx: ContextTypes.DEFAULT_TYPE):
     else:
         await ctx.bot.send_photo(chat_id=update.effective_chat.id, photo=buf, caption=caption)
 
-# ───────────────────- Plot helpers ─────────────────────────────────────────
-def _find_gaps(ts, factor: float = 2.0):
-    if len(ts) < 2:
-        return [(0, len(ts) - 1)], [], 0
-
-    intervals = [(ts[i] - ts[i - 1]).total_seconds() for i in range(1, len(ts))]
-    med = median(intervals) if intervals else 0
-    if med <= 0:
-        med = max(intervals) if intervals else 60  # fallback 60 s
-    thr = med * factor
-
-    segments, gaps = [], []
-    start = 0
-    for i, dt in enumerate(intervals, start=1):
-        if dt > thr:
-            segments.append((start, i - 1))
-            gaps.append((ts[i - 1], ts[i]))
-            start = i
-    segments.append((start, len(ts) - 1))
-    return segments, gaps, thr
-
-
-TIME_RE = re.compile(r"^(\d+)([smhd])$", re.I)
-
-def parse_timespan(tokens: list[str]) -> int:
-    """Parse duration tokens like ['2d', '3h', '15m'] into seconds."""
-    total = 0
-    for t in tokens:
-        m = TIME_RE.match(t)
-        if not m:
-            raise ValueError(f"invalid time token: {t}")
-        val = int(m.group(1))
-        unit = m.group(2).lower()
-        if unit == "s":
-            total += val
-        elif unit == "m":
-            total += val * 60
-        elif unit == "h":
-            total += val * 3600
-        elif unit == "d":
-            total += val * 86400
-    return total
-
-
-def _plot_segments(ax, ts, ys, segments, *args, **kwargs):
-    first = True
-    col = None
-    for s, e in segments:
-        if first:
-            line, = ax.plot(ts[s:e+1], ys[s:e+1], *args, **kwargs)
-            col = line.get_color()
-            first = False
-        else:
-            kw = dict(kwargs)
-            kw.pop("label", None)
-            kw["color"] = col
-            ax.plot(ts[s:e+1], ys[s:e+1], *args, **kw)
-def _make_figure(seconds: int):
-    """Return figure sized proportionally to the requested time span."""
-    days = seconds / 86_400
-    width = 12
-    if days >= 1:
-        width *= 1.5 if days < 1.5 else days
-    dpi = 500
-    fig, ax = plt.subplots(figsize=(width, 6), dpi=dpi)
-
-    base = 9
-    plt.rcParams.update({
-        "font.size": base,
-        "axes.titlesize": base + 4,
-        "axes.labelsize": base + 2,
-        "xtick.labelsize": base - 1,
-        "ytick.labelsize": base - 1,
-        "legend.fontsize": base - 1,
-    })
-    return fig, ax
-def plot_metric(secret: str, metric: str, seconds: int):
-    rows = fetch_metrics(secret, int(time.time()) - seconds)
-    if not rows:
-        return None
-
-    ts = [datetime.fromtimestamp(r[0]) for r in rows]
-
-    idx_map = {
-        "cpu": (1, "CPU %", "%", (0, 100)),
-        "ram": (2, "RAM %", "%", (0, 100)),
-        "gpu": (3, "GPU %", "%", (0, 100)),
-        "vram": (4, "VRAM %", "%", (0, 100)),
-        "net_up": (5, "Net Up", "bit/s", None),
-        "net_down": (6, "Net Down", "bit/s", None),
-    }
-    col_idx, label, ylab, ylim = idx_map[metric]
-    ys = [np.nan if rows[i][col_idx] is None else rows[i][col_idx] for i in range(len(rows))]
-
-    if metric.startswith("net_"):
-        max_val = max([v for v in ys if not np.isnan(v)] or [0])
-        scale, unit = best_unit(max_val)
-        if scale != 1:
-            ys = [v / scale if not np.isnan(v) else np.nan for v in ys]
-        ylab = unit
-        if max_val:
-            ylim = (0, (max_val / scale) * 1.1)
-
-    segments, gaps, _ = _find_gaps(ts)
-
-    plt.style.use("dark_background")
-    fig, ax = _make_figure(seconds)
-
-    _plot_segments(ax, ts, ys, segments, linewidth=1.5)
-
-    for g0, g1 in gaps:
-        ax.axvspan(g0, g1, facecolor="none", hatch="//", edgecolor="white", alpha=0.3, linewidth=0)
-
-    ax.set_title(f"{label} за {timedelta(seconds=seconds)}")
-    ax.set_xlabel("Время")
-    ax.set_ylabel(ylab)
-    if ylim:
-        ax.set_ylim(*ylim)
-    ax.grid(True, linestyle="--", linewidth=0.3)
-    fig.autofmt_xdate()
-
-    buf = io.BytesIO()
-    plt.tight_layout()
-    fig.savefig(buf, dpi=fig.dpi, format="png")
-    plt.close(fig)
-    buf.seek(0)
-    return buf
-
-def plot_net(secret: str, seconds: int):
-    rows = fetch_metrics(secret, int(time.time()) - seconds)
-    if not rows:
-        return None
-
-    ts = [datetime.fromtimestamp(r[0]) for r in rows]
-    up = [np.nan if r[5] is None else r[5] for r in rows]
-    down = [np.nan if r[6] is None else r[6] for r in rows]
-    max_val = max(
-        [v for v in up + down if not np.isnan(v)] or [0]
-    )
-    scale, unit = best_unit(max_val)
-    if scale != 1:
-        up = [v / scale if not np.isnan(v) else np.nan for v in up]
-        down = [v / scale if not np.isnan(v) else np.nan for v in down]
-    ylim_top = None
-    if max_val:
-        ylim_top = (max_val / scale) * 1.1
-    segments, gaps, _ = _find_gaps(ts)
-
-    plt.style.use("dark_background")
-    fig, ax = _make_figure(seconds)
-
-    _plot_segments(ax, ts, up, segments, label="Up", linewidth=1.2)
-    _plot_segments(ax, ts, down, segments, label="Down", linewidth=1.2)
-
-    for g0, g1 in gaps:
-        ax.axvspan(g0, g1, facecolor="none", hatch="//", edgecolor="white", alpha=0.3, linewidth=0)
-
-    ax.set_title(f"Net за {timedelta(seconds=seconds)}")
-    ax.set_xlabel("Время")
-    ax.set_ylabel(unit)
-    if ylim_top is not None:
-        ax.set_ylim(0, ylim_top)
-    ax.grid(True, linestyle="--", linewidth=0.3)
-    ax.legend(loc="upper left", fontsize="small")
-    fig.autofmt_xdate()
-
-    buf = io.BytesIO()
-    plt.tight_layout()
-    fig.savefig(buf, dpi=fig.dpi, format="png")
-    plt.close(fig)
-    buf.seek(0)
-    return buf
-def plot_all_metrics(secret: str, seconds: int):
-    rows = fetch_metrics(secret, int(time.time()) - seconds)
-    if not rows:
-        return None
-
-    ts = [datetime.fromtimestamp(r[0]) for r in rows]
-    segments, gaps, _ = _find_gaps(ts)
-
-    cpu = [r[1] for r in rows]
-    ram = [r[2] for r in rows]
-    gpu = [np.nan if r[3] is None else r[3] for r in rows]
-    vram = [np.nan if r[4] is None else r[4] for r in rows]
-
-    plt.style.use("dark_background")
-
-    fig, ax = _make_figure(seconds)
-
-
-    for ys, lab in ((cpu, "CPU %"), (ram, "RAM %")):
-        _plot_segments(ax, ts, ys, segments, label=lab, linewidth=1.2)
-    if not all(np.isnan(g) for g in gpu):
-        _plot_segments(ax, ts, gpu, segments, label="GPU %", linewidth=1.2)
-    if not all(np.isnan(v) for v in vram):
-        _plot_segments(ax, ts, vram, segments, label="VRAM %", linewidth=1.2)
-
-    for g0, g1 in gaps:
-        ax.axvspan(g0, g1, facecolor="none", hatch="//", edgecolor="white", alpha=0.3, linewidth=0)
-
-    ax.set_ylim(0, 100)
-    ax.set_title(f"Все метрики за {timedelta(seconds=seconds)}")
-    ax.set_xlabel("Время")
-    ax.set_ylabel("%")
-    ax.grid(True, linestyle="--", linewidth=0.3)
-    ax.legend(loc="upper left", fontsize="small")
-    fig.autofmt_xdate()
-
-    buf = io.BytesIO()
-    plt.tight_layout()
-    fig.savefig(buf, dpi=fig.dpi, format="png")
-    plt.close(fig)
-    buf.seek(0)
-    return buf
-
-
-def fetch_metrics_full(secret: str, since: int) -> list[dict]:
-    rows = sql.execute(
-        "SELECT ts, cpu, ram, gpu, vram, net_up, net_down, ram_used, ram_total, vram_used, vram_total FROM metrics "
-        "WHERE secret=? AND ts>=? ORDER BY ts ASC",
-        (secret, since),
-    ).fetchall()
-    if not rows:
-        return []
-
-    def avg(vals):
-        vals = [v for v in vals if v is not None]
-        return sum(vals) / len(vals) if vals else None
-
-    grouped = []
-    chunk = []
-    for r in rows:
-        chunk.append(r)
-        if len(chunk) == 6:
-            grouped.append(_avg_chunk_full(chunk, avg))
-            chunk = []
-    if chunk:
-        grouped.append(_avg_chunk_full(chunk, avg))
-    return grouped
-
-
-def _avg_chunk_full(chunk, avg_fn):
-    r_last = chunk[-1]
-    return {
-        "ts": r_last[0],
-        "cpu": avg_fn([r[1] for r in chunk]),
-        "ram": avg_fn([r[2] for r in chunk]),
-        "gpu": avg_fn([r[3] for r in chunk]),
-        "vram": avg_fn([r[4] for r in chunk]),
-        "net_up": avg_fn([r[5] for r in chunk]),
-        "net_down": avg_fn([r[6] for r in chunk]),
-        "ram_used": avg_fn([r[7] for r in chunk]),
-        "ram_total": r_last[8],
-        "vram_used": avg_fn([r[9] for r in chunk]),
-        "vram_total": r_last[10],
-    }
-
-
-MEM_UNITS = {
-    "b": 1,
-    "kb": 1024,
-    "kib": 1024,
-    "mb": 1024 ** 2,
-    "mib": 1024 ** 2,
-    "gb": 1024 ** 3,
-    "gib": 1024 ** 3,
-}
-
-NET_UNITS = {
-    "bit": 1,
-    "kbit": 1000,
-    "mbit": 1000 ** 2,
-    "gbit": 1000 ** 3,
-}
-
-
-def plot_custom(secret: str, metrics: list[str], seconds: int, ylim_top: float | None, unit: str | None):
-    rows = fetch_metrics_full(secret, int(time.time()) - seconds)
-    if not rows:
-        return None
-
-    ts = [datetime.fromtimestamp(r["ts"]) for r in rows]
-    segments, gaps, _ = _find_gaps(ts)
-
-    plt.style.use("dark_background")
-    fig, ax = _make_figure(seconds)
-
-    unit_category = None
-    ylab = unit
-
-    data_sets: list[tuple[list[float], str]] = []
-
-    for m in metrics:
-        m = m.lower()
-        if m == "net":
-            metrics.extend(["net_up", "net_down"])
-            continue
-
-    for m in metrics:
-        m = m.lower()
-        label = m.upper()
-        ys = []
-        cat = ""
-        default_unit = "%"
-
-        if m == "cpu":
-            ys = [r["cpu"] for r in rows]
-            cat = "percent"
-        elif m == "gpu":
-            ys = [r["gpu"] for r in rows]
-            cat = "percent"
-        elif m == "ram":
-            if unit and unit.lower() not in {"%", "percent"}:
-                factor = MEM_UNITS.get(unit.lower(), 1024 ** 2)
-                ys = [r["ram_used"] / factor if r["ram_used"] is not None else np.nan for r in rows]
-                cat = "bytes"
-                default_unit = unit
-            else:
-                ys = [r["ram"] for r in rows]
-                cat = "percent"
-        elif m == "vram":
-            if unit and unit.lower() not in {"%", "percent"}:
-                factor = MEM_UNITS.get(unit.lower(), 1024 ** 2)
-                ys = [r["vram_used"] / factor if r["vram_used"] is not None else np.nan for r in rows]
-                cat = "bytes"
-                default_unit = unit
-            else:
-                ys = [r["vram"] for r in rows]
-                cat = "percent"
-        elif m == "net_up":
-            cat = "net"
-            default_unit = "bit/s"
-            if unit and unit.lower() in NET_UNITS:
-                scale = NET_UNITS[unit.lower()]
-                ys = [r["net_up"] * 8 / scale if r["net_up"] is not None else np.nan for r in rows]
-            else:
-                max_val = max([r["net_up"] or 0 for r in rows])
-                scale, auto_unit = best_unit(max_val)
-                ys = [r["net_up"] / scale if r["net_up"] is not None else np.nan for r in rows]
-                if not unit:
-                    ylab = auto_unit
-            label = "Up"
-        elif m == "net_down":
-            cat = "net"
-            default_unit = "bit/s"
-            if unit and unit.lower() in NET_UNITS:
-                scale = NET_UNITS[unit.lower()]
-                ys = [r["net_down"] * 8 / scale if r["net_down"] is not None else np.nan for r in rows]
-            else:
-                max_val = max([r["net_down"] or 0 for r in rows])
-                scale, auto_unit = best_unit(max_val)
-                ys = [r["net_down"] / scale if r["net_down"] is not None else np.nan for r in rows]
-                if not unit:
-                    ylab = auto_unit
-            label = "Down"
-        else:
-            continue
-
-        if unit_category is None:
-            unit_category = cat
-            if not ylab:
-                ylab = default_unit
-        elif unit_category != cat:
-            raise ValueError("Единицы метрик не совпадают")
-
-        data_sets.append((ys, label))
-
-    for ys, lab in data_sets:
-        _plot_segments(ax, ts, ys, segments, label=lab, linewidth=1.2)
-
-    for g0, g1 in gaps:
-        ax.axvspan(g0, g1, facecolor="none", hatch="//", edgecolor="white", alpha=0.3, linewidth=0)
-
-    ax.set_xlabel("Время")
-    ax.set_ylabel(ylab or "%")
-    if ylim_top is not None:
-        ax.set_ylim(0, ylim_top)
-    ax.set_title(f"{'/'.join([l for _, l in data_sets])} за {timedelta(seconds=seconds)}")
-    ax.grid(True, linestyle="--", linewidth=0.3)
-    if len(data_sets) > 1:
-        ax.legend(loc="upper left", fontsize="small")
-    fig.autofmt_xdate()
-
-    buf = io.BytesIO()
-    plt.tight_layout()
-    fig.savefig(buf, dpi=fig.dpi, format="png")
-    plt.close(fig)
-    buf.seek(0)
-    return buf
 
 # ─────────────────────- Callback handler ───────────────────────────────────
 async def cb_action(update: Update, ctx: ContextTypes.DEFAULT_TYPE):
